Update SecurityConfig and login default page

+++ src/main/java/com/portfolio/portfolioapp/repository/ProjectRepository.java
@@ -1,10 +1,10 @@
 package com.portfolio.portfolioapp.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import com.portfolio.portfolioapp.entity.Project;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProjectRepository extends JpaRepository<Project, Integer> {
 
-    long countByStatus(String status);
+    Integer countByStatus(String status);
 
 }
